+++ controller.py
@@ -31,7 +31,7 @@ class Controller(nn.Module):
     def forward(self, states, goals):
         # preprocess input data
         x = torch.unsqueeze(states, dim=1) - torch.unsqueeze(states, dim=0)
-        x = torch.concat([x, torch.unsqueeze(torch.eye(x.size(dim=0)), dim=2)], dim=2)
+        x = torch.concat([x, torch.unsqueeze(torch.eye(x.size(dim=0)).type_as(x), dim=2)], dim=2)
         x, _ = remove_distant_agents(x, TOP_K)
         
         # build local observation mask
